requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pandas==2.2.2
duckdb==0.7.1
requests==2.31.0
beautifulsoup4==4.12.2
matplotlib==3.8.0
numpy==1.26.4
python-dotenv==1.0.0
Pillow==10.1.0
openai==1.8.5
